test(app): separate epoch change tests
