+++ core/application/src/tests/epoch_change.rs
@@ -1,0 +1,184 @@
+use tempfile::tempdir;
+
+use super::*;
+
+#[tokio::test]
+async fn test_epoch_changed() {
+    let temp_dir = tempdir().unwrap();
+
+    // Create a genesis committee and seed the application state with it.
+    let committee_size = 4;
+    let (committee, keystore) = create_genesis_committee(committee_size);
+    let (update_socket, query_runner) = test_init_app(&temp_dir, committee);
+    let required_signals = calculate_required_signals(committee_size);
+
+    let epoch = 0;
+    let nonce = 1;
+
+    // Have (required_signals - 1) say they are ready to change epoch
+    // make sure the epoch doesn't change each time someone signals
+    for node in keystore.iter().take(required_signals - 1) {
+        // Make sure epoch didn't change
+        let res = change_epoch!(&update_socket, &node.node_secret_key, nonce, epoch);
+        assert!(!res.change_epoch);
+    }
+    // check that the current epoch is still 0
+    assert_eq!(query_runner.get_epoch_info().epoch, 0);
+
+    // Have the last needed committee member signal the epoch change and make sure it changes
+    let res = change_epoch!(
+        &update_socket,
+        &keystore[required_signals].node_secret_key,
+        nonce,
+        epoch
+    );
+    assert!(res.change_epoch);
+
+    // Query epoch info and make sure it incremented to new epoch
+    assert_eq!(query_runner.get_epoch_info().epoch, 1);
+}
+
+#[tokio::test]
+async fn test_change_epoch_reverts_account_key() {
+    let temp_dir = tempdir().unwrap();
+
+    let committee_size = 4;
+    let (committee, _keystore) = create_genesis_committee(committee_size);
+    let (update_socket, _query_runner) = test_init_app(&temp_dir, committee);
+
+    // Account Secret Key
+    let secret_key = AccountOwnerSecretKey::generate();
+
+    let change_epoch = UpdateMethod::ChangeEpoch { epoch: 0 };
+
+    let update = prepare_update_request_account(change_epoch, &secret_key, 1);
+    expect_tx_revert!(update, &update_socket, ExecutionError::OnlyNode);
+}
+
+#[tokio::test]
+async fn test_change_epoch_reverts_node_does_not_exist() {
+    let temp_dir = tempdir().unwrap();
+
+    // Create a genesis committee and seed the application state with it.
+    let committee_size = 4;
+    let (committee, _keystore) = create_genesis_committee(committee_size);
+    let (update_socket, _query_runner) = test_init_app(&temp_dir, committee);
+
+    // Unknown Node Key (without Stake)
+    let node_secret_key = NodeSecretKey::generate();
+    let change_epoch = UpdateMethod::ChangeEpoch { epoch: 0 };
+
+    let update = prepare_update_request_node(change_epoch, &node_secret_key, 1);
+    expect_tx_revert!(update, &update_socket, ExecutionError::NodeDoesNotExist);
+}
+
+#[tokio::test]
+async fn test_change_epoch_reverts_insufficient_stake() {
+    let temp_dir = tempdir().unwrap();
+
+    // Create a genesis committee and seed the application state with it.
+    let committee_size = 4;
+    let (committee, _keystore) = create_genesis_committee(committee_size);
+    let (update_socket, query_runner) = test_init_app(&temp_dir, committee);
+
+    let owner_secret_key = AccountOwnerSecretKey::generate();
+    // New Node key
+    let node_secret_key = NodeSecretKey::generate();
+
+    // Stake less than the minimum required amount.
+    let minimum_stake_amount: HpUfixed<18> = query_runner.get_staking_amount().into();
+    let less_than_minimum_stake_amount: HpUfixed<18> =
+        minimum_stake_amount / HpUfixed::<18>::from(2u16);
+    deposit_and_stake!(
+        &update_socket,
+        &owner_secret_key,
+        1,
+        &less_than_minimum_stake_amount,
+        &node_secret_key.to_pk(),
+        [0; 96].into()
+    );
+
+    let change_epoch = UpdateMethod::ChangeEpoch { epoch: 0 };
+    let update = prepare_update_request_node(change_epoch, &node_secret_key, 1);
+    expect_tx_revert!(update, &update_socket, ExecutionError::InsufficientStake);
+}
+
+#[tokio::test]
+async fn test_epoch_change_reverts_epoch_already_changed() {
+    let temp_dir = tempdir().unwrap();
+
+    // Create a genesis committee and seed the application state with it.
+    let committee_size = 4;
+    let (committee, keystore) = create_genesis_committee(committee_size);
+    let (update_socket, query_runner) = test_init_app(&temp_dir, committee);
+
+    // call epoch change
+    simple_epoch_change!(&update_socket, &keystore, &query_runner, 0);
+    assert_eq!(query_runner.get_epoch_info().epoch, 1);
+
+    let change_epoch = UpdateMethod::ChangeEpoch { epoch: 0 };
+    let update = prepare_update_request_node(change_epoch, &keystore[0].node_secret_key, 2);
+    expect_tx_revert!(update, &update_socket, ExecutionError::EpochAlreadyChanged);
+}
+
+#[tokio::test]
+async fn test_epoch_change_reverts_epoch_has_not_started() {
+    let temp_dir = tempdir().unwrap();
+
+    // Create a genesis committee and seed the application state with it.
+    let committee_size = 4;
+    let (committee, keystore) = create_genesis_committee(committee_size);
+    let (update_socket, _query_runner) = test_init_app(&temp_dir, committee);
+
+    let change_epoch = UpdateMethod::ChangeEpoch { epoch: 1 };
+    let update = prepare_update_request_node(change_epoch, &keystore[0].node_secret_key, 1);
+    expect_tx_revert!(update, &update_socket, ExecutionError::EpochHasNotStarted);
+}
+
+#[tokio::test]
+async fn test_epoch_change_reverts_not_committee_member() {
+    let temp_dir = tempdir().unwrap();
+
+    // Create a genesis committee and seed the application state with it.
+    let committee_size = 4;
+    let (committee, _keystore) = create_genesis_committee(committee_size);
+    let (update_socket, query_runner) = test_init_app(&temp_dir, committee);
+
+    let owner_secret_key = AccountOwnerSecretKey::generate();
+    // New Node key
+    let node_secret_key = NodeSecretKey::generate();
+
+    // Stake less than the minimum required amount.
+    let minimum_stake_amount: HpUfixed<18> = query_runner.get_staking_amount().into();
+
+    deposit_and_stake!(
+        &update_socket,
+        &owner_secret_key,
+        1,
+        &minimum_stake_amount,
+        &node_secret_key.to_pk(),
+        [0; 96].into()
+    );
+
+    let change_epoch = UpdateMethod::ChangeEpoch { epoch: 0 };
+    let update = prepare_update_request_node(change_epoch, &node_secret_key, 1);
+    expect_tx_revert!(update, &update_socket, ExecutionError::NotCommitteeMember);
+}
+
+#[tokio::test]
+async fn test_epoch_change_reverts_already_signaled() {
+    let temp_dir = tempdir().unwrap();
+
+    // Create a genesis committee and seed the application state with it.
+    let committee_size = 4;
+    let (committee, keystore) = create_genesis_committee(committee_size);
+    let (update_socket, _query_runner) = test_init_app(&temp_dir, committee);
+
+    let change_epoch = UpdateMethod::ChangeEpoch { epoch: 0 };
+    let update = prepare_update_request_node(change_epoch.clone(), &keystore[0].node_secret_key, 1);
+    expect_tx_success!(update, &update_socket);
+
+    // Second update
+    let update = prepare_update_request_node(change_epoch, &keystore[0].node_secret_key, 2);
+    expect_tx_revert!(update, &update_socket, ExecutionError::AlreadySignaled);
+}
